+++ src/baseconfig.h
@@ -20,6 +20,7 @@ class BaseConfig {
     const String&   GetMqttRoot()      const {return mqtt_root;}
     const String&   GetMqttBasePath()  const {return mqtt_basepath;}
     const bool&     UseRandomMQTTClientID() const { return mqtt_UseRandomClientID; }
+    const bool&     GetUseETH()        const { return useETH; }
     const uint8_t&  GetDebugLevel()    const {return debuglevel;}
      
   private:
@@ -30,6 +31,7 @@ class BaseConfig {
     String    mqtt_root;
     String    mqtt_basepath;
     bool      mqtt_UseRandomClientID;
+    bool      useETH;  // otherwise use WIFI
     uint8_t   debuglevel;
 
 };