+++ src/modbus.h
@@ -10,7 +10,7 @@
 #include <ArduinoQueue.h>
 #include <HardwareSerial.h>
 
-#define DEBUGMODE
+//#define DEBUGMODE
 
 class modbus {
 
@@ -64,6 +64,9 @@ class modbus {
     uint8_t                 pin_RX;               // Serial Receive pin
     uint8_t                 pin_TX;               // Serial Transmit pin
     uint8_t                 pin_RTS;              // Direction control pin
+    uint8_t                 default_pin_RX;       // Serial Receive pin
+    uint8_t                 default_pin_TX;       // Serial Transmit pin
+    uint8_t                 default_pin_RTS;      // Direction control pin
     uint8_t                 ClientID;             // 0x01
     uint32_t                Baudrate;             // 19200
     uint16_t                TxIntervalLiveData;   // 5