+++ src/mqtt.cpp
@@ -1,38 +1,166 @@
 #include "mqtt.h"
 
 MQTT::MQTT(AsyncWebServer* server, DNSServer *dns, const char* MqttServer, uint16_t MqttPort, String MqttBasepath, String MqttRoot, char* APName, char* APpassword): 
-server(server), dns(dns), mqtt_root(MqttRoot), mqtt_basepath(MqttBasepath) { 
+server(server), dns(dns), mqtt_root(MqttRoot), mqtt_basepath(MqttBasepath), ConnectStatusWifi(false), ConnectStatusMqtt(false) { 
   
   this->subscriptions = new std::vector<String>{};
-
-  espClient = WiFiClient();
-  WiFi.mode(WIFI_STA); 
+ 
   this->mqtt = new PubSubClient();
-  
-  this->wifiManager = new AsyncWiFiManager(server, dns);
-
-  if (Config->GetDebugLevel() >=4) wifiManager->setDebugOutput(true); 
-    else wifiManager->setDebugOutput(false); 
-
-  wifiManager->setConnectTimeout(60);
-  wifiManager->setConfigPortalTimeout(300);
   WiFi.setHostname(this->mqtt_root.c_str());
-  Serial.println("WiFi Start");
-  
-//  wifiManager->startConfigPortal("OnDemandAP");
+  WiFi.onEvent(std::bind(&MQTT::WifiOnEvent, this, std::placeholders::_1));
 
-  if (!wifiManager->autoConnect(APName, APpassword) ) {
-    Serial.println("failed to connect and start configPortal");
-    wifiManager->startConfigPortal(APName, APpassword);
+  if (Config->GetDebugLevel() >=3) {
+    Serial.printf("Go into %s Mode", (Config->GetUseETH()?"ETH":"Wifi"));
+  }
+
+  if (Config->GetUseETH()) {
+    //ETH.begin(1, 16, 23, 18, ETH_PHY_LAN8720, ETH_CLOCK_GPIO0_IN);
+    ETH.begin(this->lan_shields[0].PHY_ADDR,
+              this->lan_shields[0].PHY_POWER,
+              this->lan_shields[0].PHY_MDC,
+              this->lan_shields[0].PHY_MDIO,
+              this->lan_shields[0].PHY_TYPE,
+              this->lan_shields[0].CLK_MODE);
+
+    this->ETH_waitForConnect();
+  } else {
+
+    WiFi.mode(WIFI_STA);
+    this->wifiManager = new AsyncWiFiManager(server, dns);
+
+    if (Config->GetDebugLevel() >=4) wifiManager->setDebugOutput(true); 
+      else wifiManager->setDebugOutput(false); 
+
+    wifiManager->setConnectTimeout(60);
+    wifiManager->setConfigPortalTimeout(300);
+    Serial.println("WiFi Start");
+    
+  //  wifiManager->startConfigPortal("OnDemandAP");
+
+    if (!wifiManager->autoConnect(APName, APpassword) ) {
+      Serial.println("failed to connect and start configPortal");
+      wifiManager->startConfigPortal(APName, APpassword);
+    }
   }
   
-  Serial.printf("WiFi connected with local IP: %s\n", WiFi.localIP().toString().c_str());
   //WiFi.printDiag(Serial);
 
   Serial.printf("Starting MQTT (%s:%d)\n", MqttServer, MqttPort);
+  espClient = WiFiClient();
+  
   this->mqtt->setClient(espClient);
   this->mqtt->setServer(MqttServer, MqttPort);
   this->mqtt->setCallback([this] (char* topic, byte* payload, unsigned int length) { this->callback(topic, payload, length); });
+}
+
+void MQTT::WifiOnEvent(WiFiEvent_t event) {
+    if (Config->GetDebugLevel()>=4) {Serial.printf("[WiFi-event] event: %d\n", event);}
+
+    switch (event) {
+        case ARDUINO_EVENT_WIFI_READY: 
+            Serial.println("WiFi interface ready");
+            break;
+        case ARDUINO_EVENT_WIFI_SCAN_DONE:
+            Serial.println("Completed scan for access points");
+            break;
+        case ARDUINO_EVENT_WIFI_STA_START:
+            Serial.println("WiFi client started");
+            break;
+        case ARDUINO_EVENT_WIFI_STA_STOP:
+            Serial.println("WiFi clients stopped");
+            break;
+        case ARDUINO_EVENT_WIFI_STA_CONNECTED:
+            Serial.println("Connected to access point");
+            break;
+        case ARDUINO_EVENT_WIFI_STA_DISCONNECTED:
+            Serial.println("Disconnected from WiFi access point");
+            this->ConnectStatusWifi = false;
+            break;
+        case ARDUINO_EVENT_WIFI_STA_AUTHMODE_CHANGE:
+            Serial.println("Authentication mode of access point has changed");
+            break;
+        case ARDUINO_EVENT_WIFI_STA_GOT_IP:
+            Serial.printf("WiFi connected with local IP: %s\n", WiFi.localIP().toString().c_str());
+            this->ipadresse = WiFi.localIP();
+            this->ConnectStatusWifi = true;
+            break;
+        case ARDUINO_EVENT_WIFI_STA_LOST_IP:
+            Serial.println("Lost IP address and IP address is reset to 0");
+            this->ConnectStatusWifi = false;
+            this->ipadresse = (0,0,0,0);
+            break;
+        case ARDUINO_EVENT_WPS_ER_SUCCESS:
+            Serial.println("WiFi Protected Setup (WPS): succeeded in enrollee mode");
+            break;
+        case ARDUINO_EVENT_WPS_ER_FAILED:
+            Serial.println("WiFi Protected Setup (WPS): failed in enrollee mode");
+            break;
+        case ARDUINO_EVENT_WPS_ER_TIMEOUT:
+            Serial.println("WiFi Protected Setup (WPS): timeout in enrollee mode");
+            break;
+        case ARDUINO_EVENT_WPS_ER_PIN:
+            Serial.println("WiFi Protected Setup (WPS): pin code in enrollee mode");
+            break;
+        case ARDUINO_EVENT_WIFI_AP_START:
+            Serial.println("WiFi access point started");
+            break;
+        case ARDUINO_EVENT_WIFI_AP_STOP:
+            Serial.println("WiFi access point  stopped");
+            break;
+        case ARDUINO_EVENT_WIFI_AP_STACONNECTED:
+            Serial.println("Client connected");
+            break;
+        case ARDUINO_EVENT_WIFI_AP_STADISCONNECTED:
+            Serial.println("Client disconnected");
+            break;
+        case ARDUINO_EVENT_WIFI_AP_STAIPASSIGNED:
+            Serial.println("Assigned IP address to client");
+            break;
+        case ARDUINO_EVENT_WIFI_AP_PROBEREQRECVED:
+            Serial.println("Received probe request");
+            break;
+        case ARDUINO_EVENT_WIFI_AP_GOT_IP6:
+            Serial.println("AP IPv6 is preferred");
+            break;
+        case ARDUINO_EVENT_WIFI_STA_GOT_IP6:
+            Serial.println("STA IPv6 is preferred");
+            break;
+        case ARDUINO_EVENT_ETH_GOT_IP6:
+            Serial.println("Ethernet IPv6 is preferred");
+            break;
+        case ARDUINO_EVENT_ETH_START:
+            Serial.println("Ethernet started");
+            break;
+        case ARDUINO_EVENT_ETH_STOP:
+            Serial.println("Ethernet stopped");
+            break;
+        case ARDUINO_EVENT_ETH_CONNECTED:
+            Serial.println("Ethernet connected");
+            break;
+        case ARDUINO_EVENT_ETH_DISCONNECTED:
+            Serial.println("Ethernet disconnected");
+            this->ConnectStatusWifi = false;
+            this->ipadresse = (0,0,0,0);
+            break;
+        case ARDUINO_EVENT_ETH_GOT_IP:
+            if (!this->ConnectStatusWifi) {
+              Serial.printf("ETH MAC: %s, IPv4: %s, %s, Mbps: %d\n", 
+                ETH.macAddress().c_str(), 
+                ETH.localIP().toString().c_str(),
+                (ETH.fullDuplex()?"FULL_DUPLEX":"HALF_DUPLEX"),
+                ETH.linkSpeed()
+              );
+              this->ipadresse = ETH.localIP();
+              this->ConnectStatusWifi = true;
+            }
+            break;
+        default: break;
+    }
+}
+
+void MQTT::ETH_waitForConnect() {
+  while (!this->ConnectStatusWifi)
+    delay(100);
 }
 
 void MQTT::reconnect() {
@@ -117,10 +245,10 @@ String MQTT::getTopic(String subtopic, bool fulltopic) {
   return subtopic;
 }
 
-void MQTT::Publish_IP() {
+void MQTT::Publish_IP() { 
   char buffer[15] = {0};
   memset(&buffer[0], 0, sizeof(buffer));
-  snprintf(buffer, sizeof(buffer), "%s", WiFi.localIP().toString().c_str());
+  snprintf(buffer, sizeof(buffer), "%s", this->ipadresse.toString().c_str());
   Publish_String("IP", buffer, false);
 }
 
@@ -138,7 +266,7 @@ void MQTT::Subscribe(String topic) {
   snprintf(buffer, sizeof(buffer), "%s/#", topic.c_str());
   this->subscriptions->push_back(buffer);
   if (this->mqtt->connected()) {
-    this->mqtt->subscribe(buffer); 
+    this->mqtt->subscribe(buffer);
     if (Config->GetDebugLevel() >=3) {
       Serial.print(F("MQTT now subscribed to: ")); 
       Serial.println(buffer);
@@ -156,7 +284,7 @@ void MQTT::ClearSubscriptions() {
   this->subscriptions->shrink_to_fit();
 }
 
-void MQTT::loop() {
+void MQTT::loop() { 
   if (this->mqtt_root != Config->GetMqttRoot()) {
     if (Config->GetDebugLevel() >=3) {
       Serial.printf("MQTT DeviceName has changed via Web Configuration from %s to %s \n", this->mqtt_root.c_str(), Config->GetMqttRoot().c_str());
@@ -176,32 +304,24 @@ void MQTT::loop() {
   }
 
   // WIFI lost, try to reconnect
-  if (WiFi.status() != WL_CONNECTED) {
-    if (Config->GetDebugLevel() > 1) {
-      Serial.println("WIFI lost, try to reconnect...");
-    }
+/*  if (!Config->GetUseETH() && !this->ConnectStatusWifi) {
     wifiManager->setConfigPortalTimeout(0);
-    while (WiFi.status() != WL_CONNECTED) {
+    while (!this->ConnectStatusWifi) {
       delay(5000);
       WiFi.begin();        
     }
   }
-  
+*/
+
   // WIFI ok, MQTT lost
-  if (!this->mqtt->connected() && WiFi.status() == WL_CONNECTED) { 
+  if (!this->mqtt->connected() && this->ConnectStatusWifi) { 
     if (millis() - mqttreconnect_lasttry > 10000) {
-      espClient = WiFiClient();
+      //espClient = WiFiClient();
       this->reconnect(); 
       this->mqttreconnect_lasttry = millis();
     }
-  } else if (WiFi.status() == WL_CONNECTED) { 
+  } else if (this->ConnectStatusWifi) { 
     this->mqtt->loop();
-  }
-
-  if (WiFi.status() == WL_CONNECTED) {
-    this->ConnectStatusWifi = true;
-  } else {
-    this->ConnectStatusWifi = false;
   }
 
   if (this->mqtt->connected()) {